+++ naive_flow/tracker/checkpoint/get.py
@@ -1,7 +1,8 @@
 import re
 import os
 from .utils import list_checkpoints, get_latest_log_dir
-from ..base.arg_parser import get_args
+from ..base.arg_parser import default_arg_parser_is_used, get_args
+
 
 def parse_args():
     """Use command-line arguments to determine whether or how tracker loads checkpoints
@@ -10,15 +11,17 @@ def parse_args():
 
     """
 
-    args = get_args()
-    if args.command in ["best", "load_best"]:
-        return _best(log_dir=args.log_dir)
-    if args.command in ["latest", "load_latest"]:
-        return _latest(log_dir=args.log_dir)
-    if args.command == "load":
-        return args.checkpoint
+    if default_arg_parser_is_used():
+        args = get_args()
+        if args.command in ["best", "load_best"]:
+            return _best(log_dir=args.log_dir)
+        if args.command in ["latest", "load_latest"]:
+            return _latest(log_dir=args.log_dir)
+        if args.command == "load":
+            return args.checkpoint
 
     return None
+
 
 def _best(log_dir: str = None):
 
@@ -40,12 +43,15 @@ def _best(log_dir: str = None):
         return os.path.join(log_dir, best_checkpoint)
 
     if log_dir is None:
+
         def hook(log_root_dir: str):
             log_dir = get_latest_log_dir(log_root_dir)
             return find_best(log_dir)
+
         return hook
-    
+
     return find_best(log_dir)
+
 
 def best(*, log_dir: str = None, log_root_dir: str = None):
     """Load best saved checkpoint automatically
@@ -60,13 +66,11 @@ def best(*, log_dir: str = None, log_root_dir: str = None):
     if log_dir is not None:
         assert log_root_dir is None, "log_root_dir cannot be specified along with log_dir"
         return tem
-    
+
     assert log_root_dir is not None, "log_dir cannot be specified along with log_root_dir"
 
     return tem(log_root_dir)
 
-
-    
 
 def _latest(log_dir: str = None):
     """Load latest saved checkpoint automatically
@@ -79,17 +83,22 @@ def _latest(log_dir: str = None):
     def find_latest_checkpoint(log_dir: str):
         checkpoint_list = list_checkpoints(log_dir)
 
-        assert len(checkpoint_list) > 0, f"cannot find any checkpoint in dir: '{log_dir}'"
+        assert len(
+            checkpoint_list
+        ) > 0, f"cannot find any checkpoint in dir: '{log_dir}'"
 
         return os.path.join(log_dir, checkpoint_list[-1])
 
     if log_dir is None:
+
         def hook(log_root_dir: str):
             log_dir = get_latest_log_dir(log_root_dir)
             return find_latest_checkpoint(log_dir)
+
         return hook
 
     return find_latest_checkpoint(log_dir)
+
 
 def latest(*, log_dir: str = None, log_root_dir: str = None):
     """Load latest saved checkpoint automatically
@@ -104,7 +113,7 @@ def latest(*, log_dir: str = None, log_root_dir: str = None):
     if log_dir is not None:
         assert log_root_dir is None, "log_root_dir cannot be specified along with log_dir"
         return tem
-    
+
     assert log_root_dir is not None, "log_dir cannot be specified along with log_root_dir"
 
     return tem(log_root_dir)