+++ changelog.md
@@ -1,5 +1,16 @@
 # Change Log
 
+### 0.3.14
+    - Fix a bug while extra field is set, which cause `nf.strfconfig` fail
+
+### 0.3.13
+    - Checkpoint format has changed.
+    - User is allowed `nf.strfconfig(config, description='full')` to include the descriptions.
+        - as well as `nf.dump_config(..., description='inline')`
+    - `nf.tracker.get_args()` now throw error if the `arg_parser` has not been initialized.
+        - This results that the `nf.tracker.checkpoint.parse_args()` returns `None` if the `arg_parser` not initialized.
+    - make lazy loading tensorboard
+
 ### 0.3.12
     - Fix issue when metric is nan, best metric (which is nan) cannot be updated.